+++ src/App.js
@@ -1,31 +1,19 @@
-import React from 'react';
-import { Display, Keyboard, Operations } from './Components/calculator';
-import calculate from './logic/calculate';
+import { Routes, Route } from 'react-router-dom';
+import Header from './Components/header';
+import Home from './Components/home';
+import Quote from './Components/Quote';
+import Calculator from './Components/calculator';
 
-export default class App extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-    };
-  }
+const App = () => (
+  <>
+    <Header />
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path="/calculator" element={<Calculator />} />
+      <Route path="/quote" element={<Quote />} />
+      <Route path="*" element={<h1>Note Found</h1>} />
+    </Routes>
+  </>
+);
 
-  componentDidMount() {
-    this.setState(calculate({}, 'AC'));
-  }
-
-  handleClick = (e) => {
-    this.setState((prevState) => calculate(prevState, e.target.textContent));
-  };
-
-  render() {
-    return (
-      <div className="app-container">
-        <Display state={this.state} />
-        <div className="second-row">
-          <Keyboard handleClick={this.handleClick} />
-          <Operations handleClick={this.handleClick} />
-        </div>
-      </div>
-    );
-  }
-}
+export default App;
