+++ src/Components/Quote.js
@@ -1,0 +1,8 @@
+const Quote = () => (
+  <p className="Quote">
+    mathematics is not about numbers equations computations or algorithms its about understanding.
+    ~~ William Paul thurston
+  </p>
+);
+
+export default Quote;
