+++ src/Components/home.js
@@ -1,0 +1,22 @@
+const Home = () => (
+  <>
+    <h2>Welcome to our page</h2>
+    <p>
+      Lorem ipsum dolor sit amet consectetur adipisicing elit. Quo sed possimus,
+      animi, ducimus culpa sit aliquid odit sequi at qui nulla ullam dolorum odio
+      voluptate minus nobis inventore? Totam enim repellat vel, minima maiores repellendus
+      unde neque fugiat quidem at, delectus esse officia! Cumque nemo expedita molestias
+      eos numquam amet, ullam labore reiciendis vitae eius, adipisci magnam? Libero.
+    </p>
+    <br />
+    <p>
+      Lorem ipsum dolor sit amet consectetur adipisicing elit. Quo sed possimus,
+      animi, ducimus culpa sit aliquid odit sequi at qui nulla ullam dolorum odio
+      voluptate minus nobis inventore? Totam enim repellat vel, minima maiores repellendus
+      unde neque fugiat quidem at, delectus esse officia! Cumque nemo expedita molestias
+      eos numquam amet, ullam labore reiciendis vitae eius, adipisci magnam? Libero.
+    </p>
+
+  </>
+);
+export default Home;
