+++ src/index.js
@@ -1,3 +1,4 @@
+import { BrowserRouter as Router } from 'react-router-dom';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
@@ -6,4 +7,8 @@ import './App.css';
 const root = ReactDOM.createRoot(
   document.getElementById('root'),
 );
-root.render(<App />);
+root.render(
+  <Router>
+    <App />
+  </Router>,
+);
